git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5504 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/XTC/dump_xtc.h
@@ -48,7 +48,7 @@ class DumpXTC : public Dump {
   void init_style();
   int modify_param(int, char **);
   void openfile();
-  void write_header(int);
+  void write_header(bigint);
   int count();
   void pack(int *);
   void write_data(int, double *);
