+++ src/dump_local.h
@@ -55,7 +55,7 @@ class DumpLocal : public Dump {
 
   void init_style();
   int modify_param(int, char **);
-  void write_header(int);
+  void write_header(bigint);
   int count();
   void pack(int *);
   void write_data(int, double *);
